separated query options from lisst views, and organized exports
