+++ easy_task/lib/easy_task.dart
@@ -1,30 +1,43 @@
-// Entities
-export 'src/entities/assign.dart';
-export 'src/entities/group.dart';
-export 'src/entities/task.dart';
+// EASY TASK EXPORTS
 
-export 'src/functions.dart';
-export 'src/screens/assign.detail.screen.dart';
-export 'src/screens/task.create.screen.dart';
-export 'src/screens/task.detail.screen.dart';
-export 'src/screens/task.update.screen.dart';
-
+// TASK SERVICE ========
 export 'src/task.service.dart';
 
-// FOR REVIEW
+// TASK ================
+export 'src/task/task.dart';
+export 'src/task/task.query.options.dart';
+// Task Widgets
+export 'src/task/widgets/task.list_view.dart';
+// Task Screens
+export 'src/task/screens/task.create.screen.dart';
+export 'src/task/screens/task.detail.screen.dart';
+export 'src/task/screens/task.update.screen.dart';
+
+// ASSIGN ===============
+export 'src/assign/assign.dart';
+export 'src/assign/assign.query.options.dart';
+// Assign Widgets
+export 'src/assign/widgets/assign.list_view.dart';
+// Assign Screens
+export 'src/assign/screens/assign.detail.screen.dart';
+
+// GROUP ================
+export 'src/group/group.dart';
+export 'src/group/task_group.query.options.dart';
+// Group Widgets
+export 'src/group/widgets/group.list_view.dart';
+// Group Screens
+export 'src/group/screens/group.detail.screen.dart';
+
+// Others ===============
+export 'src/functions.dart';
+// Unit Test
+export 'unit_tests/todo/todo.unit_test.dart';
+
+// FOR REVIEW ===========
 // NOTE: User should not be exported
 //       Because this should be handled
 //       by the app and should be
 //       independent from this package.
 // export 'src/user.dart';
 // export 'src/defines.dart';
-
-// Widgets
-export 'src/widgets/assign.list_view.dart';
-export 'src/widgets/group.list_view.dart';
-export 'src/widgets/task.list_view.dart';
-
-// Screens
-export 'src/screens/group.detail.screen.dart';
-
-export 'unit_tests/todo/todo.unit_test.dart';
