+++ easy_task/lib/src/task/screens/task.detail.screen.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
-import 'package:easy_task/src/screens/user/user.list.screen.dart';
+import 'package:easy_task/src/assign/assign.query.options.dart';
+import 'package:easy_task/src/user/user.list.screen.dart';
 import 'package:flutter/material.dart';
 import 'package:easy_task/easy_task.dart';
