+++ easy_task/lib/src/task/screens/task.list.screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class TaskListScreen extends StatelessWidget {
+  const TaskListScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold();
+  }
+}
