+++ easy_task/pubspec.yaml
@@ -9,6 +9,7 @@ environment:
 
 dependencies:
   cloud_firestore: ^5.0.2
+  easy_helpers: 
   firebase_auth: ^5.1.1
   firebase_core: ^3.2.0
   firebase_storage: ^12.1.0
